# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/training-exercises": {
            "post": {
                "description": "Добавляет упражнение к существующей тренировке",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "training-exercises"
                ],
                "summary": "Добавить упражнение к тренировке",
                "parameters": [
                    {
                        "description": "Данные упражнения",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_EnduranNSU_trainings_internal_adapter_in_http_dto.AddExerciseToTrainingRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_EnduranNSU_trainings_internal_adapter_in_http_dto.TrainedExerciseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_EnduranNSU_trainings_internal_adapter_in_http_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_EnduranNSU_trainings_internal_adapter_in_http_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_EnduranNSU_trainings_internal_adapter_in_http_dto.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет упражнение из тренировки",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "training-exercises"
                ],
                "summary": "Удалить упражнение из тренировки",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Training ID",
                        "name": "training_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Exercise ID",
                        "name": "exercise_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_EnduranNSU_trainings_internal_adapter_in_http_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_EnduranNSU_trainings_internal_adapter_in_http_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_EnduranNSU_trainings_internal_adapter_in_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/training-exercises/{id}": {
            "put": {
                "description": "Обновляет информацию о выполненном упражнении в тренировке",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "training-exercises"
                ],
                "summary": "Обновить выполненное упражнение",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Trained Exercise ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_EnduranNSU_trainings_internal_adapter_in_http_dto.UpdateTrainedExerciseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_EnduranNSU_trainings_internal_adapter_in_http_dto.TrainedExerciseResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_EnduranNSU_trainings_internal_adapter_in_http_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_EnduranNSU_trainings_internal_adapter_in_http_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_EnduranNSU_trainings_internal_adapter_in_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/trainings/stats": {
            "get": {
                "description": "Возвращает статистику тренировок пользователя",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "trainings"
                ],
                "summary": "Получить статистику тренировок",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_EnduranNSU_trainings_internal_adapter_in_http_dto.TrainingStatsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_EnduranNSU_trainings_internal_adapter_in_http_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_EnduranNSU_trainings_internal_adapter_in_http_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_EnduranNSU_trainings_internal_adapter_in_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_EnduranNSU_trainings_internal_adapter_in_http_dto.AddExerciseToTrainingRequest": {
            "type": "object",
            "required": [
                "exercise_id",
                "training_id"
            ],
            "properties": {
                "approaches": {
                    "type": "integer",
                    "maximum": 20,
                    "minimum": 1,
                    "example": 3
                },
                "doing": {
                    "type": "string",
                    "example": "1h"
                },
                "exercise_id": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 1
                },
                "notes": {
                    "type": "string",
                    "example": "Тяжело далось"
                },
                "reps": {
                    "type": "integer",
                    "maximum": 100,
                    "minimum": 1,
                    "example": 10
                },
                "rest": {
                    "type": "string",
                    "example": "30m"
                },
                "time": {
                    "type": "string",
                    "example": "1h30m"
                },
                "training_id": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 1
                },
                "weight": {
                    "type": "number",
                    "maximum": 1000,
                    "minimum": 0,
                    "example": 50.5
                }
            }
        },
...
        "github_com_EnduranNSU_trainings_internal_adapter_in_http_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
...
        "github_com_EnduranNSU_trainings_internal_adapter_in_http_dto.TrainedExerciseResponse": {
            "type": "object",
            "properties": {
                "approaches": {
                    "type": "integer",
                    "example": 3
                },
                "doing": {
                    "type": "string",
                    "example": "1h"
                },
                "exercise_id": {
                    "type": "integer",
                    "example": 1
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "notes": {
                    "type": "string",
                    "example": "Тяжело далось"
                },
                "reps": {
                    "type": "integer",
                    "example": 10
                },
                "rest": {
                    "type": "string",
                    "example": "30m"
                },
                "time": {
                    "type": "string",
                    "example": "1h30m"
                },
                "training_id": {
                    "type": "integer",
                    "example": 1
                },
                "weight": {
                    "type": "number",
                    "example": 50.5
                }
            }
        },
...
        "github_com_EnduranNSU_trainings_internal_adapter_in_http_dto.TrainingStatsResponse": {
            "type": "object",
            "properties": {
                "average_rating": {
                    "type": "number",
                    "example": 4.5
                },
                "completed_trainings": {
                    "type": "integer",
                    "example": 12
                },
                "last_training_date": {
                    "type": "string",
                    "example": "2023-10-05T16:30:00Z"
                },
                "total_duration": {
                    "type": "string",
                    "example": "45h30m"
                },
                "total_trainings": {
                    "type": "integer",
                    "example": 15
                }
            }
        },
...
    }
}